requirements.txt
flask~=2.1.0
python-dotenv==0.20.0
google-generativeai==0.3.1
requests==2.28.1
jupyter~=1.0.0
pip~=22.2.2
panel~=0.13.1
dataclasses~=0.8
attrs~=21.4.0
pathlib~=1.0.1
toml~=0.10.2
zlib~=1.2.12
wheel~=0.37.1
brotli~=1.0.9
openssl~=1.1.1q
cryptography~=37.0.1
py~=1.11.0
lxml~=4.9.1
pillow~=9.2.0
docutils~=0.18.1
sphinx~=5.0.2
tornado~=6.1
jinja2~=3.1.4
keyring~=23.4.0
setuptools~=63.4.1
cython~=0.29.32
future~=0.18.2